chore: move app_status and appinfo_cache into util
-
