+++ src/cli/main.cpp
@@ -8,13 +8,13 @@
 #include "module/dbus_ipc/register_meta_type.h"
 #include "module/package/package.h"
 #include "module/runtime/runtime.h"
+#include "module/util/app_status.h"
 #include "module/util/command_helper.h"
 #include "module/util/env.h"
 #include "module/util/log/log_handler.h"
 #include "module/util/sysinfo.h"
 #include "module/util/xdg.h"
 #include "package_manager.h"
-#include "package_manager/impl/app_status.h"
 #include "package_manager/impl/package_manager.h"
 #include "service/impl/app_manager.h"
