+++ src/module/util/appinfo_cache.cpp
