+++ src/module/util/appinfo_cache.h
